// src/app/bi/models/ProductoOrdenBi.model.ts
export class ProductoOrdenBi {
  ordenId: number;
  producto: number;
  fecha: Date;
  descripcion: string;
  imgUrl: string;
  cantidad: number;
  precio: number;
  subtotal: number;
  mes: number;
  semana: number;
}

export class ProductoBi {
  producto: number;
  descripcion: string;
  imgUrl: string;
  cantidad: number;
  subtotal: number;
}